reorganize memory and cache
